feat: moove member management functionality with add, delete, and wage setting operations in a separate file
